+++ src/Mapper/BaseMapper.php
@@ -11,8 +11,6 @@ declare(strict_types=1);
 
 namespace Contentful\Management\Mapper;
 
-use Contentful\Core\Resource\ResourceArray;
-use Contentful\Core\Resource\ResourceInterface;
 use Contentful\Core\ResourceBuilder\MapperInterface;
 use Contentful\Core\ResourceBuilder\ObjectHydrator;
 use Contentful\Management\ResourceBuilder;
@@ -41,21 +39,5 @@ abstract class BaseMapper implements MapperInterface
     {
         $this->builder = $builder;
         $this->hydrator = new ObjectHydrator();
-    }
-
-    /**
-     * @param string|object $target either a FQCN, or an object whose class will be automatically inferred
-     * @param array         $data
-     *
-     * @return ResourceInterface|ResourceArray
-     *
-     * @deprecated 1.1 Use $this->hydrator->hydrate() instead
-     */
-    protected function hydrate($target, array $data)
-    {
-        /** @var ResourceInterface|ResourceArray $resource */
-        $resource = $this->hydrator->hydrate($target, $data);
-
-        return $resource;
     }
 }