+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased](https://github.com/contentful/contentful-management.php/compare/1.0.0...HEAD)
 
+<!-- PENDING-CHANGES -->
+**ATTENTION**: This release contains breaking changes. Please take extra care when updating to this version. See [the upgrade guide](UPGRADE-2.0.md) for more.
+
 ### Added
 
 * `DeliveryApiKey` and `PreviewApiKey` now expose the `environments` property.
@@ -12,15 +15,18 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Extensions can now be configured with installation or instance parameters, using the `Contentful\Management\Resource\Extension\Parameter` class.
 * Webhooks can now define filters, see classes defined in the `Contentful\Management\Resource\Webhook` namespace for more.
 * Webhooks now accept a transformations array, see the [documentation](https://www.contentful.com/developers/docs/references/content-management-api/#/reference/webhooks) for more.
+* Field type `Contentful\Management\Resource\ContentType\Field\RichText` was added.
 
 ### Changed
 
 * Link resolution is now delegated to the `Contentful\Management\LinkResolver` class.
 * The SDK now uses version 2 of the `contentful/core` package. **[BREAKING]**
+* System properties are no longer handled by a single class for all resources. Now each resource will have its own corresponding class. For instance, `Contentful\Management\Resource\ContentType::getSystemProperties()` will return an object of class `Contentful\Management\SystemProperties\ContentType` which will contain only appropriate methods. **[BREAKING]**
 
 ### Removed
 
 * `BaseMapper::hydrate()` has been removed. Use `$this->hydrator->hydrate()` instead. **[BREAKING]**
+<!-- /PENDING-CHANGES -->
 
 ## [1.0.0](https://github.com/contentful/contentful-management.php/tree/1.0.0) (2018-04-18)